Introduce task description models
